Crate htwdresden

Structs

Course
A course of studies.

Exam
An exam, something to study for!

Grade
A grade, hopefully a good one.

Login
A login used to authenticate with the server.

Studygroup

Enums

Degree
A degree, e.g. something you graduate with.

HTWError
A set of errors that may occur during execution.

Semester
A semester, either Winter or Summer with a corresponding year.

Type Definitions

CourseId

Year