Struct htwdresden::Exam

pub struct Exam {
    pub title: String,
    pub exam_type: String,
    pub study_branch: String,
    pub day: String,
    pub start_time: String,
    pub end_time: String,
    pub examiner: String,
    pub next_chance: String,
    pub rooms: Vec<String>,
}

An exam, something to study for!

Fields
title: String exam_type: String study_branch: String day: String start_time: String end_time: String examiner: String next_chance: String rooms: Vec<String>
Methods

impl Exam

fn for_studygroup(group: &Studygroup) -> Result<Vec<Exam>, HTWError>

Returns a list of Exams for a given study group.

Arguments

group - Studygroup

Example

use htwdresden::{Degree, Studygroup, Exam};

let group = Studygroup { year: 2016, course: 121, group: 61, degree: Degree::Bachelor };
let exams = Exam::for_studygroup(&group);

fn for_student(year: Year,
course: CourseId,
degree: Degree)
-> Result<Vec<Exam>, HTWError>

Returns a list of Exams for a given student.

Arguments

year - Year

course - Course identifier

degree - Degree

Example

use htwdresden::{Degree, Exam};

let exams = Exam::for_student(2016, 121, Degree::Bachelor);

fn for_prof(prof: &str) -> Result<Vec<Exam>, HTWError>

Returns a list of Exams for a given professor.

Arguments

prof - Professor name

Example

use htwdresden::Exam;

let exams = Exam::for_prof("Rennekamp");